requirement.txt
pytorch>=1.0.0,<1.3.0
tensorboard_logger
tqdm